Reduce code duplication in Iceberg metadata
- Combine duplicate implementation in Native and Hive
- Consolidate logic from getTableHandle
